Made seperate class for RewardValue
